tehtavat siirretty QuizPanel-komponenttiin Sectionista
